+++ ax3l/constants/DLlama.py
@@ -5,6 +5,7 @@ class DLlama:
     BIN_DIR: Final[str] = "bin"
     SERVER: Final[str] = "llama-server"
 
+    HOST: Final[str] = "0.0.0.0"
     PORT: Final[int] = 27770
     PORT_DEV: Final[int] = 27768
     PORT_QA: Final[int] = 27769